Add GlobalExceptionHandler
