+++ src/main/kotlin/uk/gov/justice/digital/hmpps/educationandworkplanapi/resource/GlobalExceptionHandler.kt
@@ -1,0 +1,73 @@
+package uk.gov.justice.digital.hmpps.educationandworkplanapi.resource
+
+import jakarta.validation.ValidationException
+import mu.KotlinLogging
+import org.springframework.http.HttpStatus.BAD_REQUEST
+import org.springframework.http.HttpStatus.INTERNAL_SERVER_ERROR
+import org.springframework.http.HttpStatus.NOT_FOUND
+import org.springframework.http.ResponseEntity
+import org.springframework.web.bind.annotation.ExceptionHandler
+import org.springframework.web.bind.annotation.RestControllerAdvice
+import uk.gov.justice.digital.hmpps.educationandworkplanapi.domain.goal.ActionPlanNotFoundException
+import uk.gov.justice.digital.hmpps.educationandworkplanapi.domain.goal.InvalidGoalException
+import uk.gov.justice.digital.hmpps.educationandworkplanapi.resource.model.ErrorResponse
+
+private val log = KotlinLogging.logger {}
+
+@RestControllerAdvice
+class GlobalExceptionHandler {
+
+  @ExceptionHandler(ActionPlanNotFoundException::class)
+  fun handleNotFoundException(e: Exception): ResponseEntity<ErrorResponse> {
+    log.info("Not found exception: {}", e.message)
+    return ResponseEntity
+      .status(NOT_FOUND)
+      .body(
+        ErrorResponse(
+          status = NOT_FOUND.value(),
+          userMessage = e.message,
+        ),
+      )
+  }
+
+  @ExceptionHandler(InvalidGoalException::class)
+  fun handleInvalidGoalException(e: Exception): ResponseEntity<ErrorResponse> {
+    log.info("Invalid goal: {}", e.message)
+    return ResponseEntity
+      .status(BAD_REQUEST)
+      .body(
+        ErrorResponse(
+          status = BAD_REQUEST.value(),
+          userMessage = e.message,
+        ),
+      )
+  }
+
+  @ExceptionHandler(ValidationException::class)
+  fun handleValidationException(e: Exception): ResponseEntity<ErrorResponse> {
+    log.info("Validation exception: {}", e.message)
+    return ResponseEntity
+      .status(BAD_REQUEST)
+      .body(
+        ErrorResponse(
+          status = BAD_REQUEST.value(),
+          userMessage = "Validation failure: ${e.message}",
+          developerMessage = e.message,
+        ),
+      )
+  }
+
+  @ExceptionHandler(java.lang.Exception::class)
+  fun handleException(e: java.lang.Exception): ResponseEntity<ErrorResponse?>? {
+    log.error("Unexpected exception", e)
+    return ResponseEntity
+      .status(INTERNAL_SERVER_ERROR)
+      .body(
+        ErrorResponse(
+          status = INTERNAL_SERVER_ERROR.value(),
+          userMessage = "Unexpected error: ${e.message}",
+          developerMessage = e.message,
+        ),
+      )
+  }
+}
